refactor(sessions): delete the now-unused legacy SessionForm
Manual session creation flows through SessionWizard (4eb4068), so the
old SessionForm component is unreachable. Drop it. The shared
useSessionFormState hook is still useful — the wizard wraps it — so
move it (and its test) into the session-wizard folder.

- delete session-form.tsx / session-form.test.tsx / index.ts
- git-mv use-session-form-state.ts into session-wizard/
- update sessions-route.test.tsx so its component mock targets the new
  module path / display string
