--- apps/web/src/features/sessions/components/session-form/index.ts
@@ -1,10 +1,0 @@
-export type {
-	CashGameFormValues,
-	RingGameOption,
-	SessionFormDefaults,
-	SessionFormValues,
-	TournamentFormValues,
-	TournamentOption,
-} from "./session-form";
-export { SessionForm } from "./session-form";
-export { useSessionFormState } from "./use-session-form-state";
